+++ api/src/config/config.ts
@@ -1,4 +1,4 @@
 export const tokensMaxAge = {
   accessToken: 10 * 1000,
-  refreshToken: 60 * 60 * 1000,
+  refreshToken: 20 * 1000,
 };
